群聊API 设计实现

+++ router/src/main/java/lyzzcw/work/zim/router/infrastructure/mapper/ImGroupMapper.java
@@ -8,7 +8,7 @@ import lyzzcw.work.zim.router.infrastructure.entity.ImGroup;
 public interface ImGroupMapper {
     int deleteByPrimaryKey(Long id);
 
-    int insert(ImGroup record);
+    Long insert(ImGroup record);
 
     int insertSelective(ImGroup record);
